+++ ThirdPartyAdapters/inmobi/inmobi/src/main/java/com/google/ads/mediation/inmobi/InMobiInitializer.java
@@ -113,7 +113,7 @@ public class InMobiInitializer {
         });
   }
 
-  interface Listener {
+  public interface Listener {
 
     /**
      * Called when the InMobi SDK initializes successfully.