+++ ThirdPartyAdapters/liftoffmonetize/liftoffmonetize/src/main/java/com/google/ads/mediation/vungle/VungleMediationAdapter.java
@@ -57,7 +57,7 @@ import java.util.HashSet;
 import java.util.List;
 
 /**
- * Mediation network adapter for Vungle.
+ * Mediation network adapter for Liftoff Monetize.
  */
 public class VungleMediationAdapter extends RtbAdapter
     implements MediationRewardedAd, RewardedAdListener {
@@ -83,7 +83,7 @@ public class VungleMediationAdapter extends RtbAdapter
   private MediationRewardedAdCallback mediationRewardedAdCallback;
 
   /**
-   * Vungle adapter error domain.
+   * Liftoff Monetize adapter error domain.
    */
   public static final String ERROR_DOMAIN = "com.google.ads.mediation.vungle";
 
@@ -98,12 +98,12 @@ public class VungleMediationAdapter extends RtbAdapter
   public static final int ERROR_INVALID_SERVER_PARAMETERS = 101;
 
   /**
-   * The requested ad size does not match a Vungle supported banner size.
+   * The requested ad size does not match a Liftoff Monetize supported banner size.
    */
   public static final int ERROR_BANNER_SIZE_MISMATCH = 102;
 
   /**
-   * Vungle requires an {@link android.app.Activity} context to request ads.
+   * Liftoff Monetize requires an {@link android.app.Activity} context to request ads.
    */
   public static final int ERROR_REQUIRES_ACTIVITY_CONTEXT = 103;
 
@@ -124,7 +124,7 @@ public class VungleMediationAdapter extends RtbAdapter
   public static final int ERROR_CANNOT_PLAY_AD = 107;
 
   /**
-   * Convert the given Vungle error into the appropriate custom error code.
+   * Convert the given Liftoff Monetize exception into the appropriate custom error code.
    */
   @NonNull
   public static AdError getAdError(@NonNull VungleError error) {
@@ -257,7 +257,8 @@ public class VungleMediationAdapter extends RtbAdapter
     placement = PlacementFinder.findPlacement(mediationExtras, serverParameters);
     if (TextUtils.isEmpty(placement)) {
       AdError error = new AdError(ERROR_INVALID_SERVER_PARAMETERS,
-          "Failed to load ad from Vungle. Missing or invalid Placement ID.", ERROR_DOMAIN);
+          "Failed to load ad from Liftoff Monetize. Missing or invalid Placement ID.",
+          ERROR_DOMAIN);
       Log.w(TAG, error.toString());
       mediationAdLoadCallback.onFailure(error);
       return;
@@ -266,7 +267,7 @@ public class VungleMediationAdapter extends RtbAdapter
     String appID = serverParameters.getString(KEY_APP_ID);
     if (TextUtils.isEmpty(appID)) {
       AdError error = new AdError(ERROR_INVALID_SERVER_PARAMETERS,
-          "Failed to load ad from Vungle. Missing or Invalid App ID.", ERROR_DOMAIN);
+          "Failed to load ad from Liftoff Monetize. Missing or Invalid App ID.", ERROR_DOMAIN);
       Log.w(TAG, error.toString());
       mediationAdLoadCallback.onFailure(error);
       return;
@@ -309,6 +310,9 @@ public class VungleMediationAdapter extends RtbAdapter
     rewardedAd.play();
   }
 
+  /**
+   * {@link RewardedAdListener} implementation from Liftoff Monetize.
+   */
   @Override
   public void onAdLoaded(@NonNull BaseAd baseAd) {
     if (mediationAdLoadCallback != null) {
@@ -376,7 +380,8 @@ public class VungleMediationAdapter extends RtbAdapter
   }
 
   /**
-   * This class is used to map Vungle rewarded video ad rewards to Google Mobile Ads SDK rewards.
+   * This class is used to map Liftoff Monetize rewarded video ad rewards to Google Mobile Ads SDK
+   * rewards.
    */
   public static class VungleReward implements RewardItem {
 
@@ -417,10 +422,10 @@ public class VungleMediationAdapter extends RtbAdapter
       @NonNull MediationRewardedAdConfiguration mediationRewardedAdConfiguration,
       @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback> callback) {
     Log.d(TAG, "loadRewardedInterstitialAd()...");
-    Log.d(TAG, "Vungle adapter was asked to load a rewarded interstitial ad. "
+    Log.d(TAG, "Liftoff Monetize adapter was asked to load a rewarded interstitial ad. "
         + "Using the rewarded ad request flow to load the ad to attempt to load a "
-        + "rewarded interstitial ad from Vungle.");
-    // Vungle Rewarded Interstitial ads use the same Rewarded Video API.
+        + "rewarded interstitial ad from Liftoff Monetize.");
+    // Liftoff Monetize Rewarded Interstitial ads use the same Rewarded Video API.
     loadRewardedAd(mediationRewardedAdConfiguration, callback);
   }
 
@@ -476,9 +481,9 @@ public class VungleMediationAdapter extends RtbAdapter
       @NonNull MediationRewardedAdConfiguration adConfiguration,
       @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback> callback) {
     Log.d(TAG, "loadRtbRewardedInterstitialAd()...");
-    Log.d(TAG, "Vungle adapter was asked to load a rewarded interstitial ad. "
+    Log.d(TAG, "Liftoff Monetize adapter was asked to load a rewarded interstitial ad. "
         + "Using the rewarded ad request flow to load the ad to attempt to load a "
-        + "rewarded interstitial ad from Vungle.");
+        + "rewarded interstitial ad from Liftoff Monetize.");
     VungleInitializer.getInstance()
         .updateCoppaStatus(adConfiguration.taggedForChildDirectedTreatment());
     // Vungle Rewarded Interstitial ads use the same Rewarded Video API.