+++ ThirdPartyAdapters/pangle/pangle/src/main/java/com/google/ads/mediation/pangle/PangleConstants.java
@@ -24,6 +24,7 @@ public class PangleConstants {
 
   public static final String PLACEMENT_ID = "placementid";
   public static final String APP_ID = "appid";
+  // The adapter error domain.
   public static final String ERROR_DOMAIN = "com.google.ads.mediation.pangle";
   public static final String PANGLE_SDK_ERROR_DOMAIN = "com.pangle.ads";