+++ ThirdPartyAdapters/dtexchange/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## DT Exchange Android Mediation Adapter Changelog
 
-#### Version 8.2.3.0 (In progress)
+#### Version 8.2.3.0
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.
 
@@ -9,7 +9,7 @@ Built and tested with:
 - DT Exchange SDK version 8.2.3.
 
 #### Version 8.2.2.1
-- Added support for passing `muteAudio` key in mediation extra to mute interstitial ads.
+- Added support for passing `muteVideo` key in mediation extra to mute interstitial ads.
 
 Built and tested with:
 - Google Mobile Ads SDK version 21.5.0.