+++ ThirdPartyAdapters/imobile/CHANGELOG.md
@@ -1,10 +1,14 @@
 ## i-mobile Android Mediation Adapter Changelog
 
-### Next Version
+### Version 2.3.1.1
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.
 
-#### Version 2.3.1.0 (in progress)
+Built and tested with:
+- Google Mobile Ads SDK version 22.0.0.
+- IMobile SDK version 2.3.1.
+
+#### Version 2.3.1.0
 - Verified compatibility with i-mobile SDK v2.3.1.
 - Updated the minimum required Google Mobile Ads SDK version to 21.5.0.