+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## InMobi Android Mediation Adapter Changelog
 
-### Version 10.5.4.1 (In progress)
+### Version 10.5.4.1
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.