+++ ThirdPartyAdapters/liftoffmonetize/CHANGELOG.md
@@ -1,12 +1,19 @@
-## Vungle Android Mediation Adapter Changelog
+## Liftoff Monetize Android Mediation Adapter Changelog
 
 #### Version 7.0.0.0
+- Rebranded adapter name from Vungle to Liftoff Monetize.
+
+Built and tested with:
+- Google Mobile Ads SDK version 22.0.0.
+- Vungle SDK version 7.0.0.
+
+#### Version 6.12.1.1
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.
 
 Built and tested with:
 - Google Mobile Ads SDK version 22.0.0.
-- Vungle SDK version 7.0.0.
+- Vungle SDK version 6.12.1.
 
 #### Version 6.12.1.0
 - Verified compatibility with Vungle SDK 6.12.1.