+++ ThirdPartyAdapters/mintegral/CHANGELOG.md
@@ -1,6 +1,13 @@
 ## Mintegral Android Mediation Adapter Changelog
 
-### Version 16.4.21.0 (In progress)
+#### Version 16.4.31.0 (In Progress)
+- Verified compatibility with Mintegral SDK 16.4.31.
+
+Built and tested with:
+- Google Mobile Ads SDK version 22.0.0.
+- Mintegral SDK version 16.4.31.
+
+#### Version 16.4.21.0
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.
 - Verified compatibility with Mintegral SDK 16.4.21.