+++ ThirdPartyAdapters/mytarget/CHANGELOG.md
@@ -1,8 +1,12 @@
 ## myTarget Android Mediation Adapter Changelog
 
-### Next Version
+### Version 5.16.4.1
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 22.0.0.
+- MyTarget SDK version 5.16.4.
 
 #### Version 5.16.4.0
 - Verified compatibility with myTarget SDK version 5.16.4.