+++ ThirdPartyAdapters/nend/CHANGELOG.md
@@ -1,8 +1,13 @@
 ## nend Android Mediation Adapter Changelog
 
-### Next Version
+#### Version 9.0.1.0
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.
+- Verified compatibility with nend SDK 9.0.1.
+
+Built and tested with
+- Google Mobile Ads SDK version 22.0.0.
+- Nend SDK version 9.0.1.
 
 #### Version 9.0.0.0
 - Verified compatibility with nend SDK 9.0.0.