+++ ThirdPartyAdapters/unity/CHANGELOG.md
@@ -1,8 +1,12 @@
 ## Unity Ads Android Mediation Adapter Changelog
 
-### Next Version
+### Version 4.6.1.1
 - Updated adapter to use new `VersionInfo` class.
 - Updated the minimum required Google Mobile Ads SDK version to 22.0.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 22.0.0.
+- Unity Ads SDK version 4.6.1.
 
 #### Version 4.6.1.0
 - Verified compatibility with Unity Ads SDK 4.6.1.